Updated the project panels
1. Now the project panels take on non-fixed size when screen is small.
This allows nice stacking when only have one column.

2. Change color of title to orange. Seems to go well with the grey
background.

3. Fixed Jia wei and her group member's names. They were combined with
each other wrongly.

TO FIX:
1. Maybe the photo video buttons on the project title can be moved to the
bottom of the panel so that it doesn't squeeze the title text

2. Some aprostrophe-s (e.g. 's) were not parsed properly by either the
database or the scraper. This should be fixed before steps

3. Most names start with a space. This is likely due to the scraper error.
Should be fixed to allow for consistency and more importanly, ensure
equality when comparing names